add plot function for GMG model

+++ pywebplot/utils.py
@@ -4,9 +4,9 @@ def clamp(x):
 
 def rgb2hex(r, g, b):
     if 0 < r < 1:
-        r = r * 255
-        g = g * 255
-        b = b * 255
+        r = int(r * 255)
+        g = int(g * 255)
+        b = int(b * 255)
     return '#%02x%02x%02x' % (clamp(r), clamp(g), clamp(b))
